+++ config/config.exs
@@ -25,6 +25,32 @@ config :logger, :console,
 # Use Jason for JSON parsing in Phoenix
 config :phoenix, :json_library, Jason
 
+# Admin
+
+config :ex_admin,
+  repo: Birdiy.Repo,
+  module: BirdiyWeb,
+  modules: [
+    BirdiyWeb.Admin.Dashboard,
+    BirdiyWeb.Admin.Accounts.Admin,
+    BirdiyWeb.Admin.Accounts.User,
+    BirdiyWeb.Admin.Diy.Project,
+    BirdiyWeb.Admin.Diy.ProjectCategory,
+    BirdiyWeb.Admin.Diy.ProjectTopic,
+    BirdiyWeb.Admin.Timeline.Post,
+    BirdiyWeb.Admin.Timeline.Activity
+  ],
+  skin_color: :black
+
+config :birdiy, :pow,
+  user: Birdiy.Accounts.Admin,
+  repo: Birdiy.Repo,
+  web_module: BirdiyWeb
+
+config :gettext, default_locale: "en/us"
+
 # Import environment specific config. This must remain at the bottom
 # of this file so it overrides the configuration defined above.
 import_config "#{Mix.env()}.exs"
+
+config :xain, :after_callback, {Phoenix.HTML, :raw}