+++ lib/birdiy/accounts/user.ex
@@ -61,10 +61,13 @@ defmodule Birdiy.Accounts.User do
 
   @doc false
   def changeset(user, attrs) do
-    attrs = put_random_filename(attrs, [:image])
+    attrs =
+      attrs
+      |> put_random_filename([:image])
+      |> decode_base64_image([:image])
 
     user
-    |> cast(attrs, [:name, :image, :facebook_id, :google_id])
+    |> cast(attrs, [:name, :image, :facebook_id, :google_id, :banned_at])
     |> put_random_name()
     |> put_random_access_key()
     |> cast_attachments(attrs, [:image])