+++ lib/birdiy_web/admin/diy/project_topic.ex
@@ -1,0 +1,6 @@
+defmodule BirdiyWeb.Admin.Diy.ProjectTopic do
+  use ExAdmin.Register
+
+  register_resource Birdiy.Diy.ProjectTopic do
+  end
+end