+++ lib/birdiy_web/controllers/project_controller.ex
@@ -7,7 +7,7 @@ defmodule BirdiyWeb.ProjectController do
 
   def show(conn, %{"id" => id}) do
     case Diy.project_from_global_id(id) do
-      {:ok, project} ->
+      {:ok, %Diy.Project{deleted_at: nil} = project} ->
         project =
           Repo.preload(project, [
             :author,