+++ lib/birdiy_web/router.ex
@@ -1,5 +1,7 @@
 defmodule BirdiyWeb.Router do
   use BirdiyWeb, :router
+  use ExAdmin.Router
+  use Pow.Phoenix.Router
 
   pipeline :browser do
     plug :accepts, ["html"]
@@ -14,6 +16,11 @@ defmodule BirdiyWeb.Router do
     plug BirdiyWeb.Context
   end
 
+  pipeline :protected_admin do
+    plug Pow.Plug.RequireAuthenticated,
+      error_handler: Pow.Phoenix.PlugErrorHandler
+  end
+
   scope "/", BirdiyWeb do
     pipe_through :browser
 
@@ -22,6 +29,16 @@ defmodule BirdiyWeb.Router do
     get "/search", SearchController, :index
     get "/search/:text", SearchController, :show
     get "/project/:id", ProjectController, :show
+  end
+
+  scope "/admin" do
+    pipe_through :browser
+    pow_session_routes()
+  end
+
+  scope "/admin", ExAdmin do
+    pipe_through [:browser, :protected_admin]
+    admin_routes()
   end
 
   scope "/api" do