+++ CHANGELOG.md
@@ -6,10 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.2] - 2019-08-06
+### Added
+- Admin panel.
+- Pow as admin authentication.
+
 ## [0.0.1] - 2019-08-03
 ### Added
 - First deployment version of this project.
 - This CHANGELOG file even this is not a open source project.
 
 [Unreleased]: https://github.com/ccmikechen/Birdiy-server/compare/v0.0.1...HEAD
+[0.0.2]: https://github.com/ccmikechen/Birdiy-server/releases/tag/v0.0.2
 [0.0.1]: https://github.com/ccmikechen/Birdiy-server/releases/tag/v0.0.1